Add a way to view all finding templates, edit them, delete them, and add new ones as necessary. The current design seems too closely linked to a specific instance of a vulnerability and not as abstract as it ought to be.
Additionally...

Remove the “Is Template” Checkbox from the add/edit finding view
After we have created a template editor, there is no longer any need to flag instances of findings as templates. (Assuming the concept here is that a template is a generic representation of a finding, and not a finding in itself. Findings may be derivative instances of a template, but a finding is not a template.)

Unset “Is Template” on all existing findings
The existing templates are actual findings from tests. We should clean these up. Need to figure out a best path through this. It may be a manual process.

Make template listing sortable
It’d be nice to sort templates by their name so that we can arbitrarily order them (i.e. OWASP categories) for quick access. This currently appears to be sorted by ID in descending order.

Must run after applying this update:
./manage.py makemigrations
./manage.py migrate
./manage.py migrate_finding_templates