Add unique constraints for some model fields
Add unique constraints on the following fields:
- `Product.name`
- `Product_Type.name`
- `Tool_Type.name`
- `JIRA_Issue.jira_id`
- `Engagement.(name,target_start,product)`
- `Endpoint.(protocol,host,path,query,fragment,product)`

Now, all these objects can be uniquely identified by their names.  Some
parts of the code already relied on this fact, for example by using
`Model.objects.get(name=<...>)`.

In turn, this requires `max_length <= 1024` for the corresponding char fields
and `innodb_large_prefix` must be ON which is the default for MySQL >= 5.7.7.

See also: https://stackoverflow.com/questions/45233362/django-says-that-mysql-does-not-allow-unique-charfields-to-have-a-max-length-2