Ensure BURP report tags are not empty. Fix #1224
BURP report format in XML are creating tags that may or may not containg a value, even though an xml tag exists.

With this change, we maintain backwards compatiblity, but also perform a check to see if a VALUE exists after we validated a TAG exists, thus addressing hard to re-produce issues whenever we have empty values for empty remediation/issueBackground tags.

Fixes #1224

Signed-off-by: Doug Morato <dm@corp.io>