Introducing DataTables to Products and Findings
Separating product tags into their own column:
Product tags were added just to the right of the product name. this made very difficult to properly export as CSV or Excel in the upcoming DataTables effort, so splitting the tags which provide relevant information in their own column was the most efficient route to support DataTables export

Introducing DataTables to Products and Findings

Removing DataTables from node package as it is not working, but adding static files.

Added DataTables to Products and Findings html pages to support better sorting, filtering and exporting.