Generate TLS Certificate
The Nginx generates a TLS Certificate on every start if there is the environment variable is not set. You can add your own cetrificate by mounting it to /etc/nginx/ssl/ and setting GENERATE_TLS_CERTIFICATE to false. You can activate the TLS configuration fo nginx with the environment variable USE_TLS.

Hardening on the TLS Configuration of nginx.

On behalf of DB Systel GmbH