Add bulk risk acceptance API
This adds two REST APIs: accept_risks on an engagement or a test. This
allows a list of CVEs (or other vulnerability identifiers) along with
justification/compensating control information and the approver of that
risk acceptance. These APIs can be used to bulk-import whitelists and
other risk acceptance criteria from external sources.

Closes #1889.