TLS Improvments (#1847)
The Nginx generates a TLS Certificate on every start if there is the environment variable is not set. You can add your own certificate by mounting it to /etc/nginx/ssl/ and setting GENERATE_TLS_CERTIFICATE to false. You can activate the TLS configuration fo nginx with the environment variable USE_TLS.

Hardening on the TLS Configuration of nginx.

* Generate TLS Certificate

* add TLS1.3

* add Docker.md and link into nginx conf

* change Default Values and Docker.md