doc: setup: switch README.MD file extension to lower case
This commit fixes an issue where the file extension for
the setup/README.MD file is set to uppercase and where
a link to this file (inside DOCKER.md) is referencing
it using lower case (leading to a 404).
I Thought it was better to switch to lowercase here
instead of fixing the link as all the other file
were set to lowercase. I grep'd the doc and it shouldn't
break other references.