Social-auth: Fix call-back URLs
This commit fixes an issue where the trailing slash of callbacks URLs is removed
(DD_SOCIAL_AUTH_TRAILING_SLASH=(bool, False) and makes the login process fail as
it differs from the LOGIN_EXEMPT_URLS URLs that are allowing URLs with the trailing
slash.

This commit also generalizes the allowed call-back URLs to ```complete/``` instead
of cherry-picking the different IdPs. I'll simplify the process to add new IdPs
in the future.

Last but not least this commit also revert what has been introduced into #PR 2079
where DD_SOCIAL_AUTH_TRAILING_SLASH was set to False. I understand now that it was
bad idea to do that as it modifies the call-back URL that is sent to the IdP by
removing the trailing slash, it'll probably break existing social-auth deployments
as the allowed-call-back-url will need to be adapted IdP-side.