integration tests fixes and improvements (#2160)
Integration tests were failing randomly in Travis. Turns our this was not Travis' fault, but our own code.

This PR fixes two main problems with Defect Dojo / Integration tests.

Dedupe tests failing. This was caused by a corner cases being affected by #1953. The cache there is local per worker, so if you have multiple workers, you can have stale system settings for at most 30s. This is usualy OK, but in integration tests it's an issue. This PR contains a workaround, proper fix to follow based on #2164

StaleElementReferenceException by Selenium. This was caused by #1586. After the page is loaded, the datatables javascript applies itself to the DOM and changes it. It wraps itself around the original products table causing the element that selenium just found to go stale. This is a bit of a race condition, explaining why it sometimes works. Fixed by waiting for the datatables element being present in the DOM.

Improvements:

Check if an import was successful. This makes sure the test fails when the import was not succesful. And also it makes sure the test waits until the import has finished before importing the next scan!

Increase wait time for background dedupe. Could be that it's just very slow.

Delete existing QA Product that could be left behind by previous failing tests. Shouldn't happen inside travis, but you never know and it also helps when running local integration tests on an existing database.

Display logs if integration tests fail.

Remove some old commented code

Add some comments here and there to explain tests

Reverted changes from earlier version of this PR:

Make dedupe synchronous so it doesn't rely on celery in the background
might be useful to actually let celery perform dedupe.