Cicd improvements: reupload mitigation, overdue (#1929)
Some CI/CD related improvements I implemented during due to due usage:

    (fix) #1926 Mitigate only once on reuploads
    show overdue also in list of active/all engagements
    ci/cd:
        set target_end to to the date of the latest scan import
        use 10 day grace period before ci/cd engagement becomes overdue

The idea is that for CI/CD engagement a target_end usually doesn't make sense. By updating the target_end whenever a scan is uploaded it becomes nicely apparent which engagements are scanned regularly. We have a policy that every project(or branch) should be scanned at least weakly, so a 10 day grace period fits nicely with that.