fix chosen js dependency (#2904)
I noticed some errors in the dependabot logs and they seem to be caused by chosen being installed from git instead of the yarn/npm registry. This PR fixes that. Affects lots of files because the path of the css/js files has changed, but it's just that.

UPATE: Turns out 95% of files are extending base.html, so I included the js+css there and removed the includes in all those seperate pages.