apiv2: check related_fields value to stick w. OAv3
This commit fixes an issue with the `related_fields` query parameter on the
`/finding/` endpoint. Previously we were only checking the presence of the
parameter itself and not its value, which is not recommended by the OpenAPIv3
specification (https://github.com/OAI/OpenAPI-Specification/pull/1632/files).
It's now mandatory to specify the boolean on top of the parameter.

  * /finding/?related_fields -> don't return related fields
  * /finding/?related_fields=false -> don't return related fields
  * /finding/?related_fields=true -> return related fields