JIRA: Allow config per engagement, incl big JIRA refactor (#3200)
To allow users to have some flexibility in their JIRA setup / mapping to Defect Dojo, we'd decided to allow users to the configure a JIRA project on the engagement level. This is an optional config that will override the config set on the product above it.

While implementing it I came across some logic related to this that I wanted to centralize and make it more DRY. While doing it I realized I might as well put all the JIRA related stuff in one helper file. And clean up some of the logic, improve connection handling, error handling, exception handling, avoid swallowing exceptions.

As I had to test "all things JIRA" I also fixed a couple of things regarding the "push all issues" stuff that has been haunting us after it was merged. Some other things found along the way and fixed:

#3205
#3206
#2664
#3224
#3240
#3261