Add PyJWT to requirements.txt
PyJWT 2.0.0 renamed some of the exceptions, which breaks social_core 3.4.0 that is used by social-auth-app-django 4.0.0. This causes new installations or container builds to throw a 500 error on attempting to log in.

From PyJWT 2.0.0 CHANGELOG:
`Removed `ExpiredSignature`, `InvalidAudience`, and `InvalidIssuer`. Use `ExpiredSignatureError`, `InvalidAudienceError`, and `InvalidIssuerError` instead.`

Requiring pyjwt==1.7.1 resolves the issue.