Bugfix at close_old_findings
Fixed a bug where the 'is_Mitigated' field is not updated when findings are automatically closed, which leads to them not been considered closed internally by DefectDojo (although on the UI it appears "Mitigated"). This causes these findings not to appear on metrics nor at the 'View Closed Findings' option.