Use endpoint status to determine vulnerable endpoints (#5336)
Until now the endpoint status hasn't been used to determine if an endpoint is vulnerable. With this PR an endpoint is marked as vulnerable, when it has active / verified findings and their endpoint status is not mitigated. Additionally this PR contains some tlc for the endpoint dialogues and a bugfix for editing metadata.