Kubernetes hostname changes (#5745)
This is a simple hack to deal with a real annoyance I had with leveraging DD's API for submitting within k8s and having a different public/gated web frontend.

It won't help those who already use this helm deployment model. But it seems easy enough to adjust to it.