Fix for Kubernetes tests (#6362)
Bitnami has changed the index.yaml for pre-2022 charts, see bitnami/charts#10539. The Kubernetes tests were not running anymore because of this.

This PR uses the pre-2022 index.yaml and fixes this problem. Nevertheless it shows DefectDojo's Kubernetes installation comes with old charts, that we should update from time to time.