tools: NeuVector: introducing NeuVector (REST) scan type
This commit adds the support of NeuVector
(https://github.com/neuvector/neuvector) tool for importing scan
results. Scan results can be exported via REST API in JSON format (that
is why the tool is named 'NeuVector (REST)'). There is no GUI for that
at the moment.

Scan results are just a list of issues found in packages installed in
a container or an image. Very similar to Twistlock.

NeuVector also provides compliance scan results. This is not supported
by the introduced tool.