Fix flaw in logic in edit_finding
If I understand it correctly the closing should be disabled if a special note is missing. So if note_type exists the closing should be NOT disabled. Therefore it makes more sense to have closing_disable = 1. Otherwise currently the flag should be closing_disabled == 0 because it was initialed with it.