Update upgrading.md to add steps to docker-compose to first pull the latest DefectDojo changes
Adding a step to git fetch/pull/merge the latest DefectDojo changes, prior to pulling the latest defectdojo/docker images.  This may not be clear to first-time users who are attempting to upgrade their docker-compose installation, as was the case for myself. 

Reasoning:  
Upon deploying the original quick start as defined in the readme using docker-compose and investigating the tool, there was an update and a new version available.  I followed the existing docker-compose upgrade guide and encountered many issues, before I figured it out from the hint at the "Building your local images" section for the "(Of course if you're doing this, then you know you have to update the source code first)".   

Definitely my own fault but it wasn't clearly obvious for me to first git pull the updated source, prior to pulling the latest images and rebuilding the containers.  I am thankful for backups to rollback, as running the guide without the proposed suggested "git fetch/pull/merge" additions at the beginning - the upgrade was not working.  

Though a great learning experience for myself and I am happy to have gone through it, I hope this suggested change may help future readers.

PS - DefectDojo is wonderful, thank you so much!