feat: add cvss value to finding and only overwrite values if not already set
This is because of XML Report Layout of Qualys. It has CVSS Values on concrete Finding and in the Knowledgebase.   The values in the concrete Finding are more accurate and are parsed at first. It would be much better to stick with those values and only use the ones from the Knowledgebase if CVSS values from concrete finding are missing