QualysParser: add cvss value to finding (#9256)
* feat: add cvss value to finding and only overwrite values if not already set

This is because of XML Report Layout of Qualys. It has CVSS Values on concrete Finding and in the Knowledgebase.   The values in the concrete Finding are more accurate and are parsed at first. It would be much better to stick with those values and only use the ones from the Knowledgebase if CVSS values from concrete finding are missing

* linting: flake8

* feat: add unit-test for cvss_score field

* remove unnecessary loop

Co-authored-by: Charles Neill <1749665+cneill@users.noreply.github.com>

* Update test_qualys_parser.py for linting

---------

Co-authored-by: MarianG <marian.gawron@deutschebahn.com>
Co-authored-by: Charles Neill <1749665+cneill@users.noreply.github.com>