new notifications type: SLA breach combined (per product) (#8659)
This commit introduces a new type of notifications: SLA breach combined.

The main difference is that notification is produced per product.
Original SLA breach notifications are generated for each applicable
findings. This may result in hundreds of messages (e-mail, slack or
teams messages) for large products. Such alerts are hardly manageable
and in the end not of much use.

With SLA breach combined notifications being enabled a user receives a
message per product with a list of findings which breach their SLA. It
can be summarized in the following manner:

subject: <SLA breach kind> <product type> <product>

body: <product summary> <list of findings>