Enhance awssecurityhub parser.py
First attempt at enhancing a parser and reading - ( https://defectdojo.github.io/django-DefectDojo/contributing/how-to-write-a-parser/ )

The goal is to add the resource_type and resource_id to join to create a host / endpoint.

A Finding should always have a host, but a host does not always contain an IP in the context of AWS Security Hub Findings.  This enhancement *should* (hopefully) add an Endpoint field to a Finding, reducing the number of Findings from AWS Security Hub that do not contain any Host/Endpoint data.

A Finding without an actual Host/Endpoint is to me - incomplete by design and for the structure of DefectDojo Findings.