Add ResourceId to Description
FindingId is good, but at scale if there are many of the same FindingIds across different ResourceIds then that information is not very distinguishable in the Finding.  This addition of the Resource ID is important for making a Finding actionable.  Ideally, there would be a way to add Hosts/Endpoints from ResourceIds, but this gets confusing because a security group (for example) should not be a host, although it is the Resource that would need an action carried out.