Add Resource ID to guardduty and inspector
Without ResourceId a FindingId, Region, Account, etc. is still insufficient without the underlying and actionable information containing the ResourceId that needs to be investigated.  Whether it is a Security Group (sg-id#), an Instance (i-id#), or any other kind of Resource, it's ideal for this to be included in the Description to make the Finding acitonable.