Update to avoid resource_id conflict
In the condition of the L42 if statement is true, it will already capture a resourceId and so I did not want to reuse the resource_id value and so I am updating my implementation to call description_resource_id_arn_value.  It's a little confusing, but I hope the distinction here helps avoid a conflict in the event described.