Remove join for resource_arns
At first the join was not flake8 compliant, I removed the join and now "join" is not defined.  I'm not sure how to fix this, but in theory a Finding has 1 ARN/ResourceId associated with it and I don't **think** that multiple are returned.  If there is a way to add this join in without the previous errors (Flake8 '\n' + name 'join' not defined when removing the '\n')  then I would think that is good to add here, instead of this commit to remove it.