Add Under Review for Jira mappings
Add Under Review to the Jira configuration and 'open_status'.  Some teams work within sprint cycles, so if they are integrated with Jira and their DefectDojo product, it would be nice to have 'Under Review' as an 'In-Progress' and 'Active' status / resolution, so that things currently being worked on, can be tracked.  This adds flexibility to the 'Open' 'Closed' 'False Positive' or 'Risk Accepted' by adding the 'Under Review' mapping resolution for Jira so that teams can manage vulnerabilities that are in-progress or 'under review'