Update Dockerfile.django* to use gid 4242
The gid set for the image is currently 1337.
This collides with istio, a popular k8s traffic
management, telemetry, and security tool, which
also uses gid 1337 to rewrite iptables so that
it can do its magic.

This results in DefectDojo's initializer being
unable to connect to postgresql.

Original bug issue:
github.com/DefectDojo/django-DefectDojo/issues/7532